Implementation/requirements.txt
attrdict==2.0.1
attrs==19.3.0
base58==2.0.0
certifi==2019.11.28
chardet==3.0.4
cytoolz==0.10.1
easysolc==0.1.4
eth-abi==2.1.1
eth-account==0.4.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.4
eth-rlp==0.1.2
eth-typing==2.2.1
eth-utils==1.8.4
hexbytes==0.2.0
idna==2.9
importlib-metadata==1.6.0
ipfshttpclient==0.4.12
jsonschema==3.2.0
lru-dict==1.1.6
multiaddr==0.0.9
netaddr==0.7.19
parsimonious==0.8.1
protobuf==3.11.3
py-solc==3.2.0
pycryptodome==3.9.7
pyrsistent==0.16.0
pysha3==1.0.2
requests==2.23.0
rlp==1.2.0
semantic-version==2.8.4
six==1.14.0
toolz==0.10.0
typing-extensions==3.7.4.1
urllib3==1.25.8
varint==1.0.2
web3==5.7.0
websockets==8.1
zipp==3.1.0
